ASP-5381 Reorganize agent project
