+++ lm-agent/lm_agent/config.py
@@ -1,6 +1,5 @@
-import logging
 import sys
-from enum import Enum
+import logging
 from pathlib import Path
 from typing import Optional
 
@@ -8,29 +7,13 @@ from pydantic import AnyHttpUrl, Field
 from pydantic_core import ValidationError
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
-logger = logging.getLogger("lm_agent.config")
+from lm_agent.constants import LogLevelEnum
 
 
 DEFAULT_DOTENV_PATH = Path("/etc/default/license-manager-agent")
 DEFAULT_CACHE_DIR = Path.home() / Path(".cache/license-manager")
-PRODUCT_FEATURE_RX = r"^.+?\..+$"
-ENCODING = "UTF8"
-TOOL_TIMEOUT = 6  # seconds
-
-
-class LogLevelEnum(str, Enum):
-    """
-    Log level name enforcement
-    """
-
-    DEBUG = "DEBUG"
-    INFO = "INFO"
-    WARNING = "WARNING"
-    ERROR = "ERROR"
-    CRITICAL = "CRITICAL"
-
-
-_DEFAULT_BIN_PATH = Path(__file__).parent.parent / "bin"
+DEFAULT_LOG_DIR = Path("/var/log/license-manager-agent")
+DEFAULT_BIN_PATH = Path(__file__).parent.parent / "bin"
 
 
 class Settings(BaseSettings):
@@ -41,47 +24,50 @@ class Settings(BaseSettings):
     LM2_AGENT_LOG_LEVEL=DEBUG
     """
 
+    # Deployment environment
     DEPLOY_ENV: str = "LOCAL"
 
-    # base url of an endpoint serving the licensemanager2 backend
+    # Log level
+    LOG_LEVEL: LogLevelEnum = LogLevelEnum.INFO
+
+    # Base URL of the License Manager API
     BACKEND_BASE_URL: AnyHttpUrl = Field("http://127.0.0.1:8000")
 
-    # location of the log directory
-    LOG_BASE_DIR: Optional[str] = None
+    # Location of the log directory
+    LOG_BASE_DIR: Optional[Path] = DEFAULT_LOG_DIR
 
-    # path to the license server features config file
-    LICENSE_SERVER_FEATURES_CONFIG_PATH: Optional[str] = None
+    # Token cache directory
+    CACHE_DIR: Path = DEFAULT_CACHE_DIR
 
-    # path to the binary for lmutil (needed for FlexLM licenses)
-    LMUTIL_PATH: Path = _DEFAULT_BIN_PATH
+    # Path to Slurm binaries
+    SCONTROL_PATH: Path = Path("/usr/bin/scontrol")
+    SACCTMGR_PATH: Path = Path("/usr/bin/sacctmgr")
+    SQUEUE_PATH: Path = Path("/usr/bin/squeue")
 
-    # path to the binary for rlmutil (needed for RLM licenses)
-    RLMUTIL_PATH: Path = _DEFAULT_BIN_PATH
+    # Path to the binary for lmutil (needed for FlexLM licenses)
+    LMUTIL_PATH: Path = DEFAULT_BIN_PATH / "lmutil"
 
-    # path to the binary for lstc_qrun (needed for LS-Dyna licenses)
-    LSDYNA_PATH: Path = _DEFAULT_BIN_PATH
+    # Path to the binary for rlmutil (needed for RLM licenses)
+    RLMUTIL_PATH: Path = DEFAULT_BIN_PATH / "rlmutil"
 
-    # path to the binary for lmxendutil (needed for LM-X licenses)
-    LMXENDUTIL_PATH: Path = _DEFAULT_BIN_PATH
+    # Path to the binary for lstc_qrun (needed for LS-Dyna licenses)
+    LSDYNA_PATH: Path = DEFAULT_BIN_PATH / "lstc_qrun"
 
-    # path to the binary for olixtool (needed for OLicense licenses)
-    OLIXTOOL_PATH: Path = _DEFAULT_BIN_PATH
+    # Path to the binary for lmxendutil (needed for LM-X licenses)
+    LMXENDUTIL_PATH: Path = DEFAULT_BIN_PATH / "lmxendutil"
 
-    # reservation name for reconciliation
+    # Path to the binary for olixtool (needed for OLicense licenses)
+    OLIXTOOL_PATH: Path = DEFAULT_BIN_PATH / "olixtool"
+
+    # Reservation name for reconciliation
     RESERVATION_IDENTIFIER: str = "license-manager-reservation"
 
     # License Manager user name to create reservations
     LM_USER: str = "license-manager"
 
-    # debug mode turns on certain dangerous operations
-    DEBUG: bool = False
-
-    # sentry specific settings
+    # Sentry specific settings
     SENTRY_DSN: Optional[str] = None
     SENTRY_SAMPLE_RATE: Optional[float] = Field(1.0, gt=0.0, le=1.0)
-
-    # log level
-    LOG_LEVEL: LogLevelEnum = LogLevelEnum.INFO
 
     # OIDC config for machine-to-machine security
     OIDC_DOMAIN: str
@@ -90,14 +76,17 @@ class Settings(BaseSettings):
     OIDC_CLIENT_SECRET: str
     OIDC_USE_HTTPS: bool = True
 
-    # Token cache directory
-    CACHE_DIR: Path = DEFAULT_CACHE_DIR
-
     # If set to `True`, reconcile will be triggered by Prolog/Epilog. Set to `False` to disable this.
     USE_RECONCILE_IN_PROLOG_EPILOG: bool = True
 
     # Stat interval used to report the cluster status to the API
     STAT_INTERVAL: int = 60
+
+    # Timeout for the license server binaries
+    TOOL_TIMEOUT: int = 6  # seconds
+
+    # Encoding used for decoding the output of the license server binaries
+    ENCODING: str = "utf-8"
 
     model_config = SettingsConfigDict(
         env_prefix="LM2_AGENT_",
@@ -108,18 +97,14 @@ class Settings(BaseSettings):
 def init_settings() -> Settings:
     """
     Build SETTINGS, and offer a way to gracefully fail if required settings cannot be loaded
-
-    TODO: Determine if this is true. If Settings is missing a required field, it should throw
-          an exception that causes the uvicorn process to exit gracefully (with an error code)
     """
     try:
         # ignoring call-arg because we expect the parameters to
         # be set either by the environment or the dotenv file
         return Settings()  # type: ignore[call-arg]
     except ValidationError as e:
+        logger = logging.getLogger("lm_agent.config")
         logger.error(f"Failed to load settings: {str(e)}")
-        # neither fastapi nor uvicorn appear to offer a way to do a graceful
-        # shutdown as of now, so this is what we've got.
         sys.exit(1)
