+++ lm-agent/lm_agent/constants.py
@@ -1,6 +1,18 @@
 from enum import Enum
 
 
+class LogLevelEnum(str, Enum):
+    """
+    Log level name enforcement
+    """
+
+    DEBUG = "DEBUG"
+    INFO = "INFO"
+    WARNING = "WARNING"
+    ERROR = "ERROR"
+    CRITICAL = "CRITICAL"
+
+
 class LicenseServerType(str, Enum):
     """
     Describe the supported license server types that may be used for fetching licenses from license servers.
@@ -11,3 +23,6 @@ class LicenseServerType(str, Enum):
     LMX = "lmx"
     LSDYNA = "lsdyna"
     OLICENSE = "olicense"
+
+
+PRODUCT_FEATURE_RX = r"^.+?\..+$"
