+++ lm-agent/lm_agent/parsing/flexlm.py
@@ -3,7 +3,7 @@ Parser for FlexLM
 """
 import re
 from typing import Dict, Optional
-from lm_agent.server_interfaces.license_server_interface import LicenseUsesItem
+from lm_agent.models import LicenseUsesItem
 
 
 HOSTWORD = r"[a-zA-Z0-9-]+"
