+++ lm-agent/lm_agent/workload_managers/slurm/slurmctld_prolog.py
@@ -13,11 +13,11 @@ if the exit status is anything other then 0, e.g. 1.
 import asyncio
 import sys
 
-from lm_agent.backend_utils.models import LicenseBookingRequest
+from lm_agent.models import LicenseBookingRequest
 from lm_agent.backend_utils.utils import get_cluster_configs_from_backend, make_booking_request
 from lm_agent.config import settings
 from lm_agent.logs import init_logging, logger
-from lm_agent.services.reconciliation import update_features
+from lm_agent.services.reconciliation import reconcile
 from lm_agent.workload_managers.slurm.cmd_utils import get_required_licenses_for_job
 from lm_agent.workload_managers.slurm.common import get_job_context
 
@@ -81,7 +81,7 @@ async def prolog():
         if settings.USE_RECONCILE_IN_PROLOG_EPILOG:
             # Force a reconciliation before we check the feature token availability.
             try:
-                await update_features()
+                await reconcile()
             except Exception as e:
                 logger.error(f"Failed to call reconcile with {e}")
                 sys.exit(1)
