+++ lm-agent/tests/parsing/test_flexlm.py
@@ -5,7 +5,7 @@ from pytest import mark
 
 from lm_agent.parsing.flexlm import parse, parse_feature_line, parse_usage_line
 
-from lm_agent.server_interfaces.license_server_interface import LicenseUsesItem
+from lm_agent.models import LicenseUsesItem
 
 
 @mark.parametrize(
