+++ lm-agent/tests/parsing/test_lmx.py
@@ -5,7 +5,7 @@ Test the LM-X parser
 from pytest import mark
 
 from lm_agent.parsing.lmx import parse, parse_feature_line, parse_in_use_line, parse_usage_line
-from lm_agent.server_interfaces.license_server_interface import LicenseUsesItem
+from lm_agent.models import LicenseUsesItem
 
 
 def test_parse_feature_line():
