+++ lm-agent/tests/parsing/test_lsdyna.py
@@ -3,7 +3,7 @@ Test the LS-Dyna parser
 """
 
 from lm_agent.parsing.lsdyna import parse, parse_program_line, parse_total_line, parse_usage_line
-from lm_agent.server_interfaces.license_server_interface import LicenseUsesItem
+from lm_agent.models import LicenseUsesItem
 
 
 def test_parse_program_line():
