+++ lm-agent/tests/server_interfaces/test_flexlm.py
@@ -6,7 +6,7 @@ from pytest import fixture, mark, raises
 from lm_agent.config import settings
 from lm_agent.exceptions import LicenseManagerBadServerOutput
 from lm_agent.server_interfaces.flexlm import FlexLMLicenseServer
-from lm_agent.server_interfaces.license_server_interface import LicenseReportItem
+from lm_agent.models import LicenseReportItem
 
 
 @fixture
