+++ lm-agent/tests/services/test_reconciliation.py
@@ -1,19 +1,11 @@
 from unittest import mock
 
-import pytest
-from httpx import Response
 from pytest import mark
-
-from lm_agent.exceptions import LicenseManagerBackendConnectionError, LicenseManagerEmptyReportError
-from lm_agent.server_interfaces.license_server_interface import LicenseReportItem
-from lm_agent.services.reconciliation import (
-    create_or_update_reservation,
-    reconcile,
-    update_features,
-)
+from lm_agent.models import LicenseReportItem
+from lm_agent.services.reconciliation import reconcile
 
 
-@pytest.mark.asyncio
+@mark.asyncio
 @mock.patch("lm_agent.services.reconciliation.return_formatted_squeue_out")
 @mock.patch("lm_agent.services.reconciliation.create_or_update_reservation")
 @mock.patch("lm_agent.services.reconciliation.get_all_features_cluster_values")
@@ -77,69 +69,3 @@ async def test__reconcile__success(
 
     await reconcile()
     create_or_update_reservation_mock.assert_called_with("abaqus.abaqus@flexlm:280")
-
-
-@pytest.mark.asyncio
-@mock.patch("lm_agent.services.reconciliation.report")
-async def test__update_features__report_empty(report_mock):
-    """
-    Check the correct behavior when the report is empty in reconcile.
-    """
-    report_mock.return_value = []
-    with pytest.raises(LicenseManagerEmptyReportError):
-        await update_features()
-
-
-@pytest.mark.asyncio
-@pytest.mark.respx(base_url="http://backend")
-@mock.patch("lm_agent.services.reconciliation.report")
-async def test__update_features__put_failed(
-    report_mock,
-    respx_mock,
-):
-    """
-    Check that when put /features/bulk status_code is not 200, should raise exception.
-    """
-    report_mock.return_value = [
-        LicenseReportItem(
-            feature_id=1,
-            product_feature="abaqus.abaqus",
-            total=1000,
-            used=200,
-            uses=[],
-        )
-    ]
-
-    respx_mock.put("/lm/features/bulk").mock(return_value=Response(status_code=400))
-
-    with pytest.raises(LicenseManagerBackendConnectionError):
-        await update_features()
-
-
-@mark.asyncio
-@mock.patch("lm_agent.services.reconciliation.scontrol_create_reservation")
-@mock.patch("lm_agent.services.reconciliation.scontrol_show_reservation")
-@mock.patch("lm_agent.services.reconciliation.scontrol_update_reservation")
-@mock.patch("lm_agent.services.reconciliation.scontrol_delete_reservation")
-async def test_create_or_update_reservation(delete_mock, update_mock, show_mock, create_mock):
-    """
-    Test that create_or_update_reservation:
-    - update the reservation if it exists
-    - delete the reservation if it can't update
-    - create the reservation if doesn't exist
-    """
-    # Update reservation if it exists
-    show_mock.return_value = "reservation_data"
-    await create_or_update_reservation("reservation_info")
-    update_mock.assert_called_once()
-
-    # Delete reservation if it can't update
-    show_mock.return_value = "reservation_data"
-    update_mock.return_value = False
-    await create_or_update_reservation("reservation_info")
-    delete_mock.assert_called()
-
-    # Create reservation if it doesn't exist
-    show_mock.return_value = False
-    await create_or_update_reservation("reservation_info")
-    create_mock.assert_called()
