+++ lm-agent/tests/workload_managers/slurm/test_slurmctld_epilog.py
@@ -10,13 +10,13 @@ from lm_agent.workload_managers.slurm.slurmctld_epilog import epilog
 
 @pytest.mark.asyncio
 @mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.get_job_context")
-@mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.update_features")
+@mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.reconcile")
 @mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.get_required_licenses_for_job")
 @mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.remove_job_by_slurm_job_id")
 async def test_epilog(
     remove_job_by_slurm_job_id_mock,
     get_required_licenses_for_job_mock,
-    update_features_mock,
+    reconcile_mock,
     get_job_context_mock,
 ):
     bookings_mock = mock.MagicMock()
@@ -32,7 +32,7 @@ async def test_epilog(
         "job_licenses": "test.feature@flexlm:10",
     }
     await epilog()
-    update_features_mock.assert_awaited_once()
+    reconcile_mock.assert_awaited_once()
     get_required_licenses_for_job_mock.assert_called_once_with("test.feature@flexlm:10")
     remove_job_by_slurm_job_id_mock.assert_awaited_once_with("1")
 
@@ -40,13 +40,13 @@ async def test_epilog(
 @pytest.mark.asyncio
 @mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.settings")
 @mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.get_job_context")
-@mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.update_features")
+@mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.reconcile")
 @mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.get_required_licenses_for_job")
 @mock.patch("lm_agent.workload_managers.slurm.slurmctld_epilog.remove_job_by_slurm_job_id")
 async def test_epilog_without_triggering_reconcile(
     remove_job_by_slurm_job_id_mock,
     get_required_licenses_for_job_mock,
-    update_features_mock,
+    reconcile_mock,
     get_job_context_mock,
     settings_mock,
 ):
@@ -66,6 +66,6 @@ async def test_epilog_without_triggering_reconcile(
     settings_mock.USE_RECONCILE_IN_PROLOG_EPILOG = False
 
     await epilog()
-    update_features_mock.assert_not_called()
+    reconcile_mock.assert_not_called()
     get_required_licenses_for_job_mock.assert_called_once_with("test.feature@flexlm:10")
     remove_job_by_slurm_job_id_mock.assert_awaited_once_with("1")
